rename AppThemeProvider to app_theme_provider
